[HICN-2] Added P2P confidential communication on hICN
P2P confidential communications exploit the TLS 1.3 protocol to let a consumer to
establish a secure communication on an hICN name. Currently we don't support the
consumer authentication (mutual authentication in TLS) and the 0-rtt session
establishment.

+++ libtransport/src/hicn/transport/protocols/reassembly.h
@@ -70,7 +70,7 @@ class BaseReassembly : public Reassembly {
   IndexVerificationManager *index_manager_;
   std::unordered_map<std::uint32_t, ContentObject::Ptr> received_packets_;
 
-  uint64_t index_;
+  uint32_t index_;
   std::unique_ptr<utils::MemBuf> read_buffer_;
 };
